Added new thread to remove zombie clients from queue list.  A zombie client is one which the user has requested a test, but left before the test ran. The old code would catch this after a 30 sec timeout when it tried to start the test.  This new code spawns a thread to walk through the queue list to see if the client is still there.
This required adding a new message type and test flag to the server and 
client code.  The client now indicates on the initial request if it can 
respond to these queue queries.  The server now remembers the clients old/new
status (pre 3.5.5 is old) and only sends probes to new clients.  

Code changes to the initialize_test() routine and passing parameters between
the parent and child enable this new function.

Incremented version number and changed Applet version to match the servers

RAC 7/9/09