changes to support compression of tcpdump, snaplog, and cputime files.
The configure.ac file changed due to the need to add the libz library to the linker

Test code went into web100-pcap.c and testoptions.c, but it was removed and everything
was put into the logging.c file.  

The web100srv.c file has an update to the writeMeta() routine to call it with more options
that need to be passed in to determine if compression is requested.

Note: compression is enabled by default.  The -z command line option disables this function.

RAC 9/9/09