More memory leak fixes and a couple of bug fixes.
I found an on-line reference from 2003 that indicated there was a bug in the libpcap
freecode() routine.  I appearred to be bumping into this bug, so this function is 
commented out for now in the web100-pcap.c file.  I also commented out the alldevfree()
call.  This should be revisited later, but since the child process that runs this code
terminates, it should free up any malloc'ed memory.

Possible bug fix in web100srv.c  -  When creating a new child a block of memory is
malloc'ed and later free'ed.  I noticed that some of the strings contained extraneous
characters.  The code now calls memset() to 0 out the block of memory before using it.
The extraneous characters are now gone.

Probable bug fix in web100srv.c - All clients now go through the FIFO linked list to
control the testing.  In pre3.5 versions only the single client mode operation uses the
FIFO queue, multi-client mode bypassed this queue.  The v3.5 code was modified to send all
clients through the queue so clients could wait if the server was busy.

In v3.5.10, a semaphore was added to protect the queue pointer manipulation routines (adding
and removing clients from this queue). This caused the server to hang at a semaphore wait state
instead of crashing due to pointer corruptions.  I finally trace this down to a child_sig()
call being made in the middle of a queue update.  The child_sig() routine can also update the 
queue, and this was causing the hang/crash.  The child_sig() call has been moved to after the
pointer manipulation is completed.

Also, implimented better SIGCHLD handling.  This signal is handled by a short routine that 
checks to see which process generated the signal.  If one of the pkt-pair children generated 
it, then ignore this signal, those signals are handles by waitpid() calls after each test 
completes.  SIGCHLD signals for the each test child should be handled by the main process, by
calling the child_sig() rouitne.  The main process also detects "defunct" processes and clean
them up by making repeated calls to the child_sig() function.

RAC 10/14/09