Modfications to fix bugs in multi-client mode operations.
Bumped version number to 3.6.0, with intermediate versions of 3.5.15, 3.5.16, 3.5.17 & 3.5.18
Version 3.6.0 should be a working version that doesn't crash the server and clients don't get
partial results.

Partial results:  Changed web100-pcap.c and testoptions.c to resolve this bug.  The problem 
was that the pkt-pair timing data wasn't getting delivered to the parent (testing) process.
The parent would then hang in a wait state until a SIGALRM fired.  By then it was usually
too late.  In testing with 2 clients, one wired and the other wireless, I found that the 
wireless (100+ msec RTT) would see ALRM's and failed tests, but the wired client would run
to completion.  I finally noticed that the parent was reading data from the pipe in larger chunks
than it should have.  That is, the child was writing 2 lines, but in some cases the parent got
everything in a single read.  This would hang the parent on the 2nd read.

To solve this I reworked the read() section of the code.  It now lives inside a select() call
and a for() loop.  After the 1st read, the code loops back to the select() to wait for the 
2nd line.  I also added a short 30 msec delay (using usleep()) into the web100-pcap.c file.
This went between the 2 writes.  This gives the parent time to pick up the 1st line before
looking for the 2nd.  Testing now shows this code is working correctly.

Server crashes and hangs:  This was the 2nd major problem with the multi-client code.  In fact
I fixed this 1st and then found the above problem.  To solve this I reworked the select(), read(),
and write() code to correctly handle an Interrupt (EINTR error).  These functions typically wait
for an event.  However, if an interrupt occurs, then they exit and report this using the EINTR
error code.  Previously, the code didn't handle this correctly so it would hang or proceed when
I wasn't expecting it to.  

I also reworked the SIGCHLD processing and the child_sig() routine.  This routine now handles
both pkt-pair children and test children properly.  

Finally I added in a little more error handling into the main test loop.  I now detect when there
is something in the queue, but the waiting variable says it should be empty.  I also added some
code to catch an error when the waiting and/or mclients variable went below zero.  

RAC 2/9/10