More changes to resolve bugs in the mlab distro.
From looking at the code this weekend (2.27.10) and running tests it appears
the part of the problem is that the server and/or client is timing out on 
reads/writes and then the test fails.  As a specific example the network.c
file contains the readn() function, which is called by the read_msg() function.
This routine reads data from the network and returns the data it found.  I 
earlier found that the read() call would exit if an interrupt was received so
this could cause the readn() routine to fail.  I also noticed that it could hang
forever if nothing ever arrived on the socket.  To resolve these problems I 
added a select with a timer to prevent an indefinate hang, and handled the errno=INTR
case.  This should have fixed thing, but I then found that both the server code AND
the client code use this same readxxx() functions, and the timeout for the server was
way too short for the client.  This cased the client to exit before the server sent it
the wait time signal.  (at least this happend in multi-client mode.)  The solution was
to make the time much longer (was 10 sec, now it's 600 sec).  This may need to be
revisited.

I then found a couple of bugs in the web100srv.c code.  In 1 case if the client times out
the waiting variable was decremented twice, causing the server to miss count waiting clients.
I also moved one of the test conditions to handle errors better, The server was attempting
to run tests with invalid test suite data, it now detects this condition.

Handled error and exit conditions better when a client can't get into the queue.

Handled a full queue bug that caused an extra client to enter the queue.

Improved the exit and error reporting for the command-line and java client.  

Fixed a bug in deploying the janalyze class and jar files.

incremented the version number to 3.6.1

Remaining, task --  The error messages on the Java applet have been updated to help
identify what was going on when the fault occurred.  This version needs to be 
patches with Seth's version and a new signed applet needs to be generated.

RAC 2/28/10