Update which signals are handled by cleanup.
This change removes the unconditional loop that registers the 'cleanup'
signal handler for *all* signals and registers the handler for specific
signals that perform some action in cleanup.

This change is to take advantage of the default implementations for
SIGINT, SIGTERM, and SIGTSTP and to avoid unnecessary handling of all
other signals.

The TODO on cleaup is removed and a TODO is added to remove SIGSEGV
after bug fixes.