Extend c2s server read loop by 1 second.
This change addresses a regression observed for web100clt during the c2s test.
In particular, before this change, web100clt can block while writing after the
c2s server stopped reading. This causes the rest of the test to fail.

The server implementation in NDT 3.7.0 attempted to read longer than the client
wrote. The change in NDT 4.0.0 made the c2s read loop strictly conditional on
testDuration. This resulted in a race condition with the client.