No more freezing if pcap process goes wrong (#51)
If packet capture fails to start for whatever reason, then the whole client gets hung in a read. When enough clients are hung, then all incoming clients get queued, and the queue never gets any smaller.

This adds a select() in front of the read() to ensure that the pipe becomes readable within a second. Thus, if the packet trace is unable to start, the test can still proceed without it.